pyproject.toml
[project]
name = "calculator-mcp"
version = "0.2.7"
description = "A calculator MCP server"
authors = [
    {name = "Rubens Gomes",email = "rubens.s.gomes@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10.0,<4.0.0"
dependencies = [
    "fastmcp>=3.0.0rc1",
    "pyyaml>=6.0.3,<7.0.0",
    "calculator-lib-rubens (>=0.1.5,<0.2.0)",
]

[project.scripts]
calculator-mcp = "calculator_mcp.main:main"

[project.urls]
Homepage = "https://github.com/rubensgomes/calculator-mcp/"
Documentation = "https://github.com/rubensgomes/calculator-mcp/README.md"
Repository = "https://github.com/rubensgomes/calculator-mcp/"

[tool.poetry]
packages = [{include = "calculator_mcp", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
source = ["calculator_mcp"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 90

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "pylint (>=4.0.4,<5.0.0)",
    "coverage (>=7.13.4,<8.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
]

[tool.black]
line-length = 80

[tool.isort]
line_length = 80
profile = "black"

[tool.mypy]
mypy_path = "src"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
